apps/digital-twin/pyproject.toml
[project]
name = "digital-twin"
version = "0.1.0"
description = "Computer Digital Twin — asset registry and site entity definitions"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.135.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.12.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.16.0",
    "asyncpg>=0.30.0",
    "redis>=5.3.0",
    "structlog>=25.1.0",
    "pyyaml>=6.0.0",
    "httpx>=0.28.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.0",
    "pytest-asyncio>=1.0.0",
    "ruff>=0.11.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
